+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Fixed message text pre-conversion occurring even when style was disabled
 - Fixed a bug in matching normal-type triggers introduced in v2.3.4
 - Switched double-click to select-word and added triple-click for select-all
-- Added support for CTRL+Z undo and CTRL+Y redo in text fields
+- Added support for short-term CTRL+Z undo and CTRL+Y redo in text fields
 - Fixed some edge-case issues with translatable message processing
 - Added support for restyling capturing groups from a regex-type trigger